removed unnecessary casts

+++ src_tests/ServidorAplicacao/Servicos/teacher/InsertSummaryTest.java
@@ -22,144 +22,143 @@ import Util.TipoAula;
  * @author Leonor Almeida
  * @author Sérgio Montelobo
  */
-public class InsertSummaryTest extends ServiceNeedsAuthenticationTestCase {
+public class InsertSummaryTest extends ServiceNeedsAuthenticationTestCase
+{
 
-	/**
-	 * @param testName
-	 */
-	public InsertSummaryTest(String testName) {
-		super(testName);
-	}
+    /**
+     * @param testName
+     */
+    public InsertSummaryTest(String testName)
+    {
+        super(testName);
+    }
 
-	protected String getDataSetFilePath() {
-		return "etc/datasets/servicos/teacher/testInsertSummaryDataSet.xml";
-	}
+    protected String getDataSetFilePath()
+    {
+        return "etc/datasets/servicos/teacher/testInsertSummaryDataSet.xml";
+    }
 
-	protected String getExpectedDataSetFilePath() {
-		return "etc/datasets/servicos/teacher/testExpectedInsertSummaryDataSet.xml";
-	}
+    protected String getExpectedDataSetFilePath()
+    {
+        return "etc/datasets/servicos/teacher/testExpectedInsertSummaryDataSet.xml";
+    }
 
-	protected String getNameOfServiceToBeTested() {
-		return "InsertSummary";
-	}
+    protected String getNameOfServiceToBeTested()
+    {
+        return "InsertSummary";
+    }
 
-	protected String[] getAuthenticatedAndAuthorizedUser() {
+    protected String[] getAuthenticatedAndAuthorizedUser()
+    {
 
-		String[] args = { "user", "pass", getApplication()};
-		return args;
-	}
+        String[] args = { "user", "pass", getApplication()};
+        return args;
+    }
 
-	protected String[] getAuthenticatedAndUnauthorizedUser() {
+    protected String[] getAuthenticatedAndUnauthorizedUser()
+    {
 
-		String[] args = { "julia", "pass", getApplication()};
-		return args;
-	}
+        String[] args = { "julia", "pass", getApplication()};
+        return args;
+    }
 
-	protected String[] getNotAuthenticatedUser() {
+    protected String[] getNotAuthenticatedUser()
+    {
 
-		String[] args = { "jccm", "pass", getApplication()};
-		return args;
-	}
+        String[] args = { "jccm", "pass", getApplication()};
+        return args;
+    }
 
-	protected Object[] getAuthorizeArguments() {
+    protected Object[] getAuthorizeArguments()
+    {
 
-		Integer executionCourseId = new Integer(24);
+        Integer executionCourseId = new Integer(24);
 
-		Calendar summaryDate = Calendar.getInstance();
-		summaryDate.set(Calendar.DAY_OF_MONTH, 1);
-		summaryDate.set(Calendar.MONTH, 2);
-		summaryDate.set(Calendar.YEAR, 1999);
+        Calendar summaryDate = Calendar.getInstance();
+        summaryDate.set(Calendar.DAY_OF_MONTH, 1);
+        summaryDate.set(Calendar.MONTH, 2);
+        summaryDate.set(Calendar.YEAR, 1999);
 
-		Calendar summaryHour = Calendar.getInstance();
-		summaryHour.set(Calendar.HOUR_OF_DAY, 12);
-		summaryHour.set(Calendar.MINUTE, 0);
-		summaryHour.set(Calendar.SECOND, 0);
+        Calendar summaryHour = Calendar.getInstance();
+        summaryHour.set(Calendar.HOUR_OF_DAY, 12);
+        summaryHour.set(Calendar.MINUTE, 0);
+        summaryHour.set(Calendar.SECOND, 0);
 
-		String title = "Titulo";
-		String text = "Texto do sumario";
-		Integer tipoAula = new Integer(TipoAula.DUVIDAS);
+        String title = "Titulo";
+        String text = "Texto do sumario";
+        Integer tipoAula = new Integer(TipoAula.DUVIDAS);
 
-		Object[] args =
-			{
-				executionCourseId,
-				summaryDate,
-				summaryHour,
-				tipoAula,
-				title,
-				text };
-		return args;
-	}
+        Object[] args = { executionCourseId, summaryDate, summaryHour, tipoAula, title, text };
+        return args;
+    }
 
-	protected String getApplication() {
-		return Autenticacao.EXTRANET;
-	}
+    protected String getApplication()
+    {
+        return Autenticacao.EXTRANET;
+    }
 
-	public void testSuccessfull() {
+    public void testSuccessfull()
+    {
 
-		try {
-			String[] args1 = getAuthenticatedAndAuthorizedUser();
-			IUserView userView = authenticateUser(args1);
+        try
+        {
+            String[] args1 = getAuthenticatedAndAuthorizedUser();
+            IUserView userView = authenticateUser(args1);
 
-			Object[] args2 = getAuthorizeArguments();
+            Object[] args2 = getAuthorizeArguments();
 
-			gestor.executar(userView, getNameOfServiceToBeTested(), args2);
+            gestor.executar(userView, getNameOfServiceToBeTested(), args2);
 
-			PersistenceBroker broker =
-				PersistenceBrokerFactory.defaultPersistenceBroker();
+            PersistenceBroker broker = PersistenceBrokerFactory.defaultPersistenceBroker();
 
-			Criteria criteria = new Criteria();
-			criteria.addOrderBy("lastModifiedDate", false);
-			Query queryCriteria = new QueryByCriteria(Summary.class, criteria);
-			ISummary summary =
-				(ISummary) broker.getObjectByQuery(queryCriteria);
-			broker.close();
+            Criteria criteria = new Criteria();
+            criteria.addOrderBy("lastModifiedDate", false);
+            Query queryCriteria = new QueryByCriteria(Summary.class, criteria);
+            ISummary summary = (ISummary) broker.getObjectByQuery(queryCriteria);
+            broker.close();
 
-			// verificar se o sumario foi inserido na base de dados
-			Calendar expectedSummaryDate = (Calendar) args2[1];
-			Calendar expectedSummaryHour = (Calendar) args2[2];
+            // verificar se o sumario foi inserido na base de dados
+            Calendar expectedSummaryDate = (Calendar) args2[1];
+            Calendar expectedSummaryHour = (Calendar) args2[2];
 
-			assertEquals(
-				summary.getExecutionCourse().getIdInternal(),
-				(Integer) args2[0]);
-			assertEquals(
-				summary.getSummaryDate().get(Calendar.DAY_OF_MONTH),
-				expectedSummaryDate.get(Calendar.DAY_OF_MONTH));
-			assertEquals(
-				summary.getSummaryDate().get(Calendar.MONTH),
-				expectedSummaryDate.get(Calendar.MONTH));
-			assertEquals(
-				summary.getSummaryDate().get(Calendar.YEAR),
-				expectedSummaryDate.get(Calendar.YEAR));
-			assertEquals(
-				summary.getSummaryHour().get(Calendar.HOUR_OF_DAY),
-				expectedSummaryHour.get(Calendar.HOUR_OF_DAY));
-			assertEquals(
-				summary.getSummaryHour().get(Calendar.MINUTE),
-				expectedSummaryHour.get(Calendar.MINUTE));
-			assertEquals(
-				summary.getSummaryHour().get(Calendar.SECOND),
-				expectedSummaryHour.get(Calendar.SECOND));
-			assertEquals(
-				summary.getSummaryType().getTipo(),
-				(Integer) args2[3]);
-			assertEquals(summary.getTitle(), (String) args2[4]);
-			assertEquals(summary.getSummaryText(), (String) args2[5]);
+            assertEquals(summary.getExecutionCourse().getIdInternal(), args2[0]);
+            assertEquals(
+                summary.getSummaryDate().get(Calendar.DAY_OF_MONTH),
+                expectedSummaryDate.get(Calendar.DAY_OF_MONTH));
+            assertEquals(
+                summary.getSummaryDate().get(Calendar.MONTH),
+                expectedSummaryDate.get(Calendar.MONTH));
+            assertEquals(
+                summary.getSummaryDate().get(Calendar.YEAR),
+                expectedSummaryDate.get(Calendar.YEAR));
+            assertEquals(
+                summary.getSummaryHour().get(Calendar.HOUR_OF_DAY),
+                expectedSummaryHour.get(Calendar.HOUR_OF_DAY));
+            assertEquals(
+                summary.getSummaryHour().get(Calendar.MINUTE),
+                expectedSummaryHour.get(Calendar.MINUTE));
+            assertEquals(
+                summary.getSummaryHour().get(Calendar.SECOND),
+                expectedSummaryHour.get(Calendar.SECOND));
+            assertEquals(summary.getSummaryType().getTipo(), args2[3]);
+            assertEquals(summary.getTitle(), (String) args2[4]);
+            assertEquals(summary.getSummaryText(), (String) args2[5]);
 
-			// apaga o sumario inserido para verificar se nao houve mais nenhuma alteracao da bd
-			Integer summaryId = summary.getIdInternal();
-			ISuportePersistente sp = SuportePersistenteOJB.getInstance();
-			sp.iniciarTransaccao();
-			sp.getIPersistentSummary().deleteByOID(Summary.class, summaryId);
-			sp.confirmarTransaccao();
+            // apaga o sumario inserido para verificar se nao houve mais nenhuma alteracao da bd
+            Integer summaryId = summary.getIdInternal();
+            ISuportePersistente sp = SuportePersistenteOJB.getInstance();
+            sp.iniciarTransaccao();
+            sp.getIPersistentSummary().deleteByOID(Summary.class, summaryId);
+            sp.confirmarTransaccao();
 
-			// verificar as alteracoes da bd
-			compareDataSetUsingExceptedDataSetTableColumns(getExpectedDataSetFilePath());
-		}
-		catch (FenixServiceException ex) {
-			fail("Insert the Summary of a Site" + ex);
-		}
-		catch (Exception ex) {
-			fail("Insert the Summary of a Site" + ex);
-		}
-	}
+            // verificar as alteracoes da bd
+            compareDataSetUsingExceptedDataSetTableColumns(getExpectedDataSetFilePath());
+        } catch (FenixServiceException ex)
+        {
+            fail("Insert the Summary of a Site" + ex);
+        } catch (Exception ex)
+        {
+            fail("Insert the Summary of a Site" + ex);
+        }
+    }
 }
